+++ consumer/src/main/java/io/alchevrier/consumer/MessageConsumer.java
@@ -4,5 +4,5 @@ import io.alchevrier.message.ConsumeResponse;
 import io.alchevrier.message.Topic;
 
 public interface MessageConsumer {
-    ConsumeResponse consume(Topic topic, long startOffset, long batchSize);
+    ConsumeResponse consume(Topic topic, long startOffset, int batchSize);
 }
